// code-app/src/generated/models/SdkmessageprocessingstepsModel.ts
export const Sdkmessageprocessingstepsasyncautodelete = {
  0: 'No',
  1: 'Yes'
} as const;
export type Sdkmessageprocessingstepsasyncautodelete = keyof typeof Sdkmessageprocessingstepsasyncautodelete;
export const Sdkmessageprocessingstepscanbebypassed = {
  0: 'No',
  1: 'Yes'
} as const;
export type Sdkmessageprocessingstepscanbebypassed = keyof typeof Sdkmessageprocessingstepscanbebypassed;
export const Sdkmessageprocessingstepscanusereadonlyconnection = {
  0: 'No',
  1: 'Yes'
} as const;
export type Sdkmessageprocessingstepscanusereadonlyconnection = keyof typeof Sdkmessageprocessingstepscanusereadonlyconnection;
export const Sdkmessageprocessingstepscomponentstate = {
  0: 'Published',
  1: 'Unpublished',
  2: 'Deleted',
  3: 'DeletedUnpublished'
} as const;
export type Sdkmessageprocessingstepscomponentstate = keyof typeof Sdkmessageprocessingstepscomponentstate;
export const Sdkmessageprocessingstepsenablepluginprofiler = {
  0: 'No',
  1: 'Yes'
} as const;
export type Sdkmessageprocessingstepsenablepluginprofiler = keyof typeof Sdkmessageprocessingstepsenablepluginprofiler;
export const Sdkmessageprocessingstepsinvocationsource = {
  '-1': 'Internal',
  0: 'Parent',
  1: 'Child'
} as const;
export type Sdkmessageprocessingstepsinvocationsource = keyof typeof Sdkmessageprocessingstepsinvocationsource;
export const Sdkmessageprocessingstepsismanaged = {
  0: 'Unmanaged',
  1: 'Managed'
} as const;
export type Sdkmessageprocessingstepsismanaged = keyof typeof Sdkmessageprocessingstepsismanaged;
export const Sdkmessageprocessingstepsmode = {
  0: 'Synchronous',
  1: 'Asynchronous'
} as const;
export type Sdkmessageprocessingstepsmode = keyof typeof Sdkmessageprocessingstepsmode;
export const Sdkmessageprocessingstepsstage = {
  5: 'InitialPre_operation_Forinternaluseonly_',
  10: 'Pre_validation',
  15: 'InternalPre_operationBeforeExternalPlugins_Forinternaluseonly_',
  20: 'Pre_operation',
  25: 'InternalPre_operationAfterExternalPlugins_Forinternaluseonly_',
  30: 'MainOperation_Forinternaluseonly_',
  35: 'InternalPost_operationBeforeExternalPlugins_Forinternaluseonly_',
  40: 'Post_operation',
  45: 'InternalPost_operationAfterExternalPlugins_Forinternaluseonly_',
  50: 'Post_operation_Deprecated_',
  55: 'FinalPost_operation_Forinternaluseonly_',
  80: 'Pre_Commitstagefiredbeforetransactioncommit_Forinternaluseonly_',
  90: 'Post_Commitstagefiredaftertransactioncommit_Forinternaluseonly_'
} as const;
export type Sdkmessageprocessingstepsstage = keyof typeof Sdkmessageprocessingstepsstage;
export const Sdkmessageprocessingstepsstatecode = {
  0: 'Enabled',
  1: 'Disabled'
} as const;
export type Sdkmessageprocessingstepsstatecode = keyof typeof Sdkmessageprocessingstepsstatecode;
export const Sdkmessageprocessingstepsstatuscode = {
  1: 'Enabled',
  2: 'Disabled'
} as const;
export type Sdkmessageprocessingstepsstatuscode = keyof typeof Sdkmessageprocessingstepsstatuscode;
export const Sdkmessageprocessingstepssupporteddeployment = {
  0: 'ServerOnly',
  1: 'MicrosoftDynamics365ClientforOutlookOnly',
  2: 'Both'
} as const;
export type Sdkmessageprocessingstepssupporteddeployment = keyof typeof Sdkmessageprocessingstepssupporteddeployment;

export interface SdkmessageprocessingstepsBase {
  asyncautodelete: Sdkmessageprocessingstepsasyncautodelete;
  canbebypassed: Sdkmessageprocessingstepscanbebypassed;
  canusereadonlyconnection: Sdkmessageprocessingstepscanusereadonlyconnection;
  category?: string;
  configuration?: string;
  description?: string;
  enablepluginprofiler?: Sdkmessageprocessingstepsenablepluginprofiler;
  eventexpander?: string;
  "EventHandler@odata.bind": string;
  eventhandlertypecode?: string;
  filteringattributes?: string;
  "FxExpressionId@odata.bind"?: string;
  "ImpersonatingUserId@odata.bind"?: string;
  introducedversion?: string;
  invocationsource: Sdkmessageprocessingstepsinvocationsource;
  iscustomizable: string;
  ishidden: string;
  mode: Sdkmessageprocessingstepsmode;
  name: string;
  "PluginTypeId@odata.bind": string;
  "PowerfxRuleId@odata.bind"?: string;
  rank: string;
  runtimeintegrationproperties?: string;
  "SdkMessageFilterId@odata.bind"?: string;
  "SdkMessageId@odata.bind": string;
  sdkmessageprocessingstepid: string;
  "SdkMessageProcessingStepSecureConfigId@odata.bind": string;
  stage: Sdkmessageprocessingstepsstage;
  statecode: Sdkmessageprocessingstepsstatecode;
  statuscode?: Sdkmessageprocessingstepsstatuscode;
  supporteddeployment: Sdkmessageprocessingstepssupporteddeployment;
}

export interface Sdkmessageprocessingsteps extends SdkmessageprocessingstepsBase {
  asyncautodeletename?: string;
  canbebypassedname?: string;
  canusereadonlyconnectionname: string;
  componentstate: Sdkmessageprocessingstepscomponentstate;
  createdbyname?: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame?: string;
  customizationlevel: string;
  enablepluginprofilername?: string;
  eventhandlername?: string;
  fxexpressionidname?: string;
  impersonatinguseridname?: string;
  invocationsourcename?: string;
  ismanaged: Sdkmessageprocessingstepsismanaged;
  ismanagedname?: string;
  modename?: string;
  modifiedbyname?: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame?: string;
  overwritetime: string;
  plugintypeidname?: string;
  powerfxruleidname?: string;
  sdkmessageidname?: string;
  sdkmessageprocessingstepidunique: string;
  solutionid: string;
  stagename?: string;
  statecodename?: string;
  statuscodename?: string;
  supporteddeploymentname?: string;
  supportingsolutionid?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  eventhandler?: object;
  _eventhandler_value?: string;
  fxexpressionid?: object;
  _fxexpressionid_value?: string;
  impersonatinguserid?: object;
  _impersonatinguserid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
  plugintypeid?: object;
  _plugintypeid_value?: string;
  powerfxruleid?: object;
  _powerfxruleid_value?: string;
  sdkmessagefilterid?: object;
  _sdkmessagefilterid_value?: string;
  sdkmessageid?: object;
  _sdkmessageid_value?: string;
  sdkmessageprocessingstepsecureconfigid?: object;
  _sdkmessageprocessingstepsecureconfigid_value?: string;
}